backend/requirements.txt
alpha-vantage==2.1.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
docopt==0.6.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Jsonpify==1.5.0
Flask-PyMongo==2.3.0
freeze==1.0.10
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kafka-python==2.0.1
lxml==4.5.0
MarkupSafe==1.1.1
mosestokenizer==1.1.0
multitasking==0.0.9
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
<<<<<<< HEAD
pandas==0.25.3
PySocks==1.7.1
=======
>>>>>>> 044d9f51a6268fd5ae0448ba04f61c7c8b3c79d2
openfile==0.0.7
pandas==0.25.3
pymongo==3.10.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
textblob==0.15.3
<<<<<<< HEAD
tweepy==3.8.0
=======
>>>>>>> 044d9f51a6268fd5ae0448ba04f61c7c8b3c79d2
toolwrapper==1.0.0
tweepy==3.8.0
uctools==1.2.1
urllib3==1.25.8
Werkzeug==1.0.0
yfinance==0.1.54
